Component source files are no longer cached when rendering during development.
Webpack is used to ensure that the component's files are always up to date. There is a performance overhead from webpack's bundling process, but it's a better alternative than restarting the server everytime you change a file.

The behaviour is controlled by the CACHE_COMPONENT_SOURCE setting, which defaults to `not django.conf.settings.DEBUG`